chore: reorganize the code
