+++ backend/src/main/java/io/mpj/webpigram/epigram/interactions/AnonymousInteractionService.java
@@ -1,4 +1,4 @@
-package io.mpj.webpigram.epigram.anonymous;
+package io.mpj.webpigram.epigram.interactions;
 
 import com.google.common.collect.ImmutableList;
 import io.mpj.webpigram.config.CaptchaVerificationService;
