+++ backend/src/main/java/io/mpj/webpigram/epigram/interactions/InteractionType.java
@@ -1,4 +1,4 @@
-package io.mpj.webpigram.epigram.anonymous;
+package io.mpj.webpigram.epigram.interactions;
 
 public enum InteractionType {
   UPVOTE,
